removed files.py & process.py after moving relevant functions into mccSearch.py
